+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## _Версия 1.22_
 - Обновлен метод Helper::sqlExtractNames
+- Обновлен метод Helper::stringSplitRange
 
 ## _Версия 1.21_
 - Добавлен метод Helper::exceptionToArray